+++ R/turns.R
@@ -26,6 +26,7 @@ NULL
 #' @param tokens A numeric vector of length 2 representing the number of
 #'   input and output tokens (respectively) used in this turn. Currently
 #'   only recorded for assistant turns.
+#' @param completed A POSIXct timestamp indicating when the turn completed.
 #' @export
 #' @return An S7 `Turn` object
 #' @examples
@@ -48,12 +49,14 @@ Turn <- new_class(
     text = new_property(
       class = class_character,
       getter = function(self) contents_text(self)
-    )
+    ),
+    completed = new_property(class = class_POSIXct | NULL)
   ),
   constructor = function(role,
                          contents = list(),
                          json = list(),
-                         tokens = c(0, 0)) {
+                         tokens = c(0, 0),
+                         completed = Sys.time()) {
 
    if (is.character(contents)) {
       contents <- list(ContentText(paste0(contents, collapse = "\n")))
@@ -63,7 +66,8 @@ Turn <- new_class(
       role = role,
       contents = contents,
       json = json,
-      tokens = tokens
+      tokens = tokens,
+      completed = completed
     )
   }
 )
@@ -147,7 +151,7 @@ normalize_turns <- function(turns = NULL,
   }
 
   if (!is.null(system_prompt)) {
-    system_turn <- Turn("system", system_prompt)
+    system_turn <- Turn("system", system_prompt, completed = NULL)
 
     # No turns; start with just the system prompt
     if (length(turns) == 0) {